Combine release scripts in to one
