+++ docs/releases/development.rst
@@ -18,3 +18,4 @@ Next release (in development)
 * Use coordinate bounds from the dataset when making polygons (:pr:`49`)
 * Fix a bug in :meth:`Ugrid.drop_geometry()` (:pr:`50`)
 * Relicense to BSD 3-Clause (:issue:`55`, :pr:`56`)
+* Automate most of the release process (:pr:`60`)
